+++ pricing/src/epilot/sdk.py
@@ -8,6 +8,7 @@ from .deprecated import Deprecated
 from .order_api import OrderAPI
 from .sdkconfiguration import SDKConfiguration
 from epilot import utils
+from epilot._hooks import SDKHooks
 from epilot.models import shared
 from typing import Callable, Dict, Union
 
@@ -87,6 +88,16 @@ class Epilot:
                 server_url = utils.template_url(server_url, url_params)
 
         self.sdk_configuration = SDKConfiguration(client, security, server_url, server_idx, retry_config=retry_config)
+
+        hooks = SDKHooks()
+
+        current_server_url, *_ = self.sdk_configuration.get_server_details()
+        server_url, self.sdk_configuration.client = hooks.sdk_init(current_server_url, self.sdk_configuration.client)
+        if current_server_url != server_url:
+            self.sdk_configuration.server_url = server_url
+
+        # pylint: disable=protected-access
+        self.sdk_configuration._hooks=hooks
        
         self._init_sdks()